make projectiles participate in interaction resolver
Includes a refactor to just use 'post-process'. It was cool to have frame resolution happen immediately after everyone needing it had gone, but including a variable number of projectiles in the mix meant the interaction resolver's 'active actors' count was stateful, and would need to be incorporated into rollback. Easier to leave it as-is, I think.