Start implementation of 'action buffer'
Concept is similar to the concept outlined in a previous commit message. Actions will be stored in a buffer as something that can be consumed. Each frame, the starting state defines lookback distance, then all users of this buffer (phase effects, mostly) can call 'consume' to try to get inputs out of it.